# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FRX Labs API
  version: 0.2.0
  description: FRX Labs — FRX Wallet + FRX Shield on Sui
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/execute:
    post:
      summary: Execute agent intent through firewall
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteRequest'
      responses:
        '200':
          description: Execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecuteResponse'
  /api/v1/agents:
    get:
      summary: List agents
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
    post:
      summary: Create agent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAgentResponse'
  /api/v1/agents/{id}:
    get:
      summary: Get agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
  /api/v1/agents/{id}/pause:
    post:
      summary: Pause agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
...
  /api/v1/agents/{id}/executions:
    get:
      summary: List agent executions
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Agent:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        owner_address: { type: string }
        status: { type: string, enum: [active, paused, revoked] }
        trust_score: { type: integer }
        created_at: { type: string, format: date-time }
    ExecuteRequest:
      type: object
      required: [action]
      properties:
        action: { type: string }
        asset_in: { type: string }
        asset_out: { type: string }
        amount: { type: number }
        protocol: { type: string }
        metadata: { type: object }
    ExecuteResponse:
      type: object
      properties:
        status: { type: string, enum: [approved, blocked, requires_review] }
        risk_score: { type: integer }
        execution_id: { type: string, format: uuid }
        reason: { type: string }
        policy_violation: { type: boolean }
        policy_check: { type: string }
        execution: { type: string }
    CreateAgentRequest:
      type: object
      required: [name, owner_address]
      properties:
        name: { type: string }
        owner_address: { type: string }
        trust_score: { type: integer }
    CreateAgentResponse:
      type: object
      properties:
        agent: { $ref: '#/components/schemas/Agent' }
        api_key: { type: string }
...